requirements.txt
flask
flask-cors
python-dotenv
requests
duckdb